backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
google-genai>=1.0.0
google-cloud-storage>=2.18.0
supabase>=2.9.0
websockets>=13.1
numpy>=2.0
python-dotenv>=1.0.1
pydantic>=2.9.2
Pillow>=10.4.0
httpx>=0.27.2
